requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.1.2
django-cors-middleware==1.3.1
django-extensions==2.1.3
django-heroku==0.3.1
django-nose==1.4.6
django-rest-swagger==2.2.0
djangorestframework==3.9.0
docopt==0.6.2
drf-yasg==1.11.0
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
openapi-codec==1.3.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.5
PyYAML==3.13
requests==2.20.0
ruamel.yaml==0.15.75
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24
whitenoise==4.1
